Swap float64 for float32

+++ kgcn/src/encoder/encode_test.py
@@ -42,7 +42,7 @@ def build_encoders(keyspace, uri="localhost:48555"):
                 'neighbour_data_type': lambda x: x,
                 'neighbour_value_long': lambda x: x,
                 'neighbour_value_double': lambda x: x,
-                'neighbour_value_boolean': lambda x: tf.cast(boolean.one_hot_boolean_encode(x), dtype=tf.float64),  # TODO Hacky, don't like it
+                'neighbour_value_boolean': lambda x: tf.to_double(boolean.one_hot_boolean_encode(x)),
                 'neighbour_value_date': lambda x: x,
                 'neighbour_value_string': lambda x: x}
